Stock Inventory Program
This program was created to record the input and change to stock in an
inventory. It involves many python files which must all be used in
conjunction for the program to work properly. The program is run in the
StockApplication.py file. The program is an example of Object-Orientated
Software Design.